requirements.txt
adabelief-pytorch
git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup
timm==v0.8.10dev0
torchmetrics==0.6.0
torch==2.0.0+cu118
pandas==1.5.3
torchvision==0.15.1+cu118
pillow==9.5.0
tqdm==4.65.0
matplotlib==3.7.1
PyYAML==6.0